requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
flask==2.3.3
fastapi==0.104.1
uvicorn==0.24.0